Refactor the json serializer into JsonHelpers
